writing-practice/requirements.txt
streamlit==1.32.2
python-dotenv>=1.0.0
pyyaml>=6.0.1
Pillow==10.2.0
pandas>=2.0.0
python-slugify>=8.0.0
requests==2.31.0
boto3>=1.34.0
botocore>=1.34.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pyyaml>=6.0
numpy>=1.24.0
pytest>=7.4.0